feat: RequestManager 클래스를 request 패키지로 이동하고 body 파싱 기능 추가
- webserver 패키지 -> request 패키지로 이동
- HTTP 요청의 body 파싱 기능 추가
